assistant/requirements.txt
openai~=1.55.0
python-dotenv~=1.0.1
pandas~=2.2.3
